Renamed to CarBreedDetails.
